Dedupe copy-pasted UI widgets: color picker, glass inset chrome, slider rows
- Extract ColorSwatchPicker (panels/color-swatch-picker.tsx) unifying the
  popover CompactColorPicker and the sidebar ColorPicker behind one props
  contract with a glass/card variant, and dedupe the verbatim-duplicated
  COLOR_SWATCHES preset list. Wall-paint SwatchGrid is left as is: it has
  no recent colours and uses a palette-grid + custom-colour-tile layout,
  so it is structurally a different widget.
- Add pc-glass--inset to plotcraft/tokens.css and use it at the 5 wall
  paint panel sites that repeated the same inline inset chrome (the two
  grids keep their deeper shadow / radius / padding as inline overrides).
- Extract SliderRow (panels/slider-row.tsx) and migrate DimensionRow,
  SignalRangeRow and VisionSliderRow in the item context popover to it.
  The shadcn-styled sliders and room-shapes SizeRow keep their own
  markup: different structure or pixel-level styling.

Pure visual refactor: rendered UI is pixel-identical.

Fixes #38
