+++ components/room-organizer/plotcraft/tokens.css
@@ -100,6 +100,17 @@
 }
 .pc-glass--dark { background-color: var(--pc-glass-tint-dark); }
 
+/* Recessed well inside a glass panel — dark inset chrome shared by toggle
+   rows and swatch/pattern grids. Grid sites override radius, padding and
+   shadow depth inline where they differ. */
+.pc-glass--inset {
+  background: rgba(0, 0, 0, 0.20);
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  border-radius: 8px;
+  padding: 3px;
+  box-shadow: inset 0 2px 4px rgba(0, 0, 0, 0.30);
+}
+
 /* ===========================================================
    Typography helpers (scoped — don't fight Tailwind defaults)
    =========================================================== */
